feat: implement platform-aware shortcuts for cmd/ctrl parity
- Add dynamically detected cmdModifier constant ('cmd' on macOS/Raycast, 'ctrl' on Linux/Vicinae/GNOME) to prevent Super key hijack
- Apply cmdModifier to all keyboard shortcut modifiers throughout search-dictionary, define-clipboard, and random-word commands
- Re-formatted and validated types
